requirements.txt
django>=3,<4
django-cors-headers
django-cron
django-filter
djangorestframework
djangorestframework-filters
djangorestframework-queryfields
djangorestframework-yaml
mysqlclient
pillow
django-uuslug